Move UI related types to '/ui.ts'
