+++ src/meta/geo/index.ts
@@ -2,12 +2,12 @@ export type { Bounds } from './bounds'
 export type { BurnedAreasUIOptions } from './burnedAreaSource'
 export { BurnedAreaKey, burnedAreaLayersMetadata } from './burnedAreaSource'
 export type { Layer } from './forest'
-export type { BurnedAreasOptions, ForestOptions, LayerSource, ProtectedAreasOptions } from './forest'
-export { LayerStatus } from './forest'
+export type { LayerSource } from './forest'
 export { agreementPalette, ForestKey, forestLayersMetadata, getRecipeAgreementAreaProperty } from './forest'
 export type { ForestEstimations, ForestEstimationsData } from './forestEstimations'
 export { ExtraEstimation, extraEstimationsMetadata } from './forestEstimations'
 export type { GeoStatisticsState } from './geoStatistics'
 export type { MosaicOptions, MosaicSource } from './mosaic'
 export { ProtectedAreaKey, protectedAreaLayersMetadata } from './protectedAreaSource'
-export type { MapPanel } from './ui'
+export type { BurnedAreasOptions, ForestOptions, MapPanel, ProtectedAreasOptions } from './ui'
+export { LayerStatus } from './ui'
